% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaField-class.R
\name{listDepth}
\alias{listDepth}
\alias{listClass}
\alias{listDefLanguage}
\alias{listOutput}
\alias{listHead}
\title{get max depth of a nested list}
\usage{
listDepth(this, thisdepth = 0)

listClass(this)

listDefLanguage(this)

listOutput(this)

listHead(this, n = 1L)
}
\arguments{
\item{this}{object to check}

\item{thisdepth}{current depth}
}
\value{
integer

character

character

list
}
\description{
get max depth of a nested list
}
\keyword{internal}
