% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaField-class.R
\name{MetaField}
\alias{MetaField}
\title{R6 class representing the information of one meta field}
\description{
R6 class representing the information of one meta field

R6 class representing the information of one meta field
}
\examples{

## ------------------------------------------------
## Method `MetaField$new`
## ------------------------------------------------

f <- MetaField$new(name = 'source_url', data = 'https://bi.zh.ch/bildungsplanung')
f$list
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Actual name of the field}

\item{\code{data}}{Actual data of the field}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{list}}{The data of a MetaField in list format}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MetaField$new()}}
\item \href{#method-check}{\code{MetaField$check()}}
\item \href{#method-print}{\code{MetaField$print()}}
\item \href{#method-clone}{\code{MetaField$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a MetaField object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaField$new(name, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(string) Name of the field}

\item{\code{data}}{Value of the field}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{MetaField} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{f <- MetaField$new(name = 'source_url', data = 'https://bi.zh.ch/bildungsplanung')
f$list
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check"></a>}}
\if{latex}{\out{\hypertarget{method-check}{}}}
\subsection{Method \code{check()}}{
Check values given in $data, update $depth
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaField$check(silent = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{(flag) in silent mode, the function returns a logical; in non-silent mode, the function throws errors and returns the object itself}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
flag/self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Create a text representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaField$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaField$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
