% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaInfo-class.R
\name{MetaInfo}
\alias{MetaInfo}
\title{R6 class representing the information of one meta field}
\description{
R6 class representing the information of one meta field

R6 class representing the information of one meta field
}
\examples{

## ------------------------------------------------
## Method `MetaInfo$new`
## ------------------------------------------------

m <- MetaInfo$new(title = 'Titel')
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{Title of the dataset}

\item{\code{source_name}}{Publishers of the dataset.}

\item{\code{source_url}}{URL where the dataset can be found, or URL for the publishers}

\item{\code{issued}}{Date/Time of the first publication of this dataset}

\item{\code{modified}}{Date/Time of the last modification of this dataset}

\item{\code{units}}{Labels of the datasets' values}

\item{\code{aggregate}}{Information on the aggregation of the datasets' values}

\item{\code{dim_order}}{Ordering of the dataset}

\item{\code{hierarchy}}{Sorting and hierarchical structure of the values within features}

\item{\code{labels}}{Labels of features and of values within features}

\item{\code{description}}{In depth information on the dataset, on features and on values within features}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MetaInfo$new()}}
\item \href{#method-clone}{\code{MetaInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a MetaInfo object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaInfo$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Informations to populate the meta fields}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{MetaInfo} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{m <- MetaInfo$new(title = 'Titel')
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MetaInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
