% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageString-class.R
\name{parseLangArgs}
\alias{parseLangArgs}
\title{parse language string arguments}
\usage{
parseLangArgs(...)
}
\arguments{
\item{...}{arguments}
}
\value{
list of language strings
}
\description{
parse language string arguments
}
\keyword{internal}
