% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageString-class.R
\name{LanguageString}
\alias{LanguageString}
\title{R6 Class representing a string in multiple languages}
\description{
R6 Class representing a string in multiple languages

R6 Class representing a string in multiple languages
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{list}}{The data of a LanguageString in list format}

\item{\code{json}}{The data of a LanguageString in json format}

\item{\code{xml}}{The data of a LanguageString in xml format}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LanguageString$new()}}
\item \href{#method-edit}{\code{LanguageString$edit()}}
\item \href{#method-clone}{\code{LanguageString$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a LanguageString object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LanguageString$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Strings named with two letter language code}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{LanguageString} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit"></a>}}
\if{latex}{\out{\hypertarget{method-edit}{}}}
\subsection{Method \code{edit()}}{
Edit the data of a LanguageString object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LanguageString$edit(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Strings named with two letter language code}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LanguageString$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
